2020ified and compiled

+++ src/main/kotlin/frc/team3324/robot/util/physics/Motors.kt
@@ -14,7 +14,7 @@ object Motors {
         abstract val Kv: Double
 
     }
-    class Cim(numberOfMotors: Double): Motor() {
+    class Cim(numberOfMotors: Int): Motor() {
         override val freeSpeed = rpmToRad(5310.0)
         override val freeCurrent = 2.7
         override val stallTorque = 2.42 * numberOfMotors
@@ -25,7 +25,7 @@ object Motors {
         override val Kv = freeSpeed / (12 - freeCurrent * resistance)
     }
 
-    class MiniCim(numberOfMotors: Double): Motor() {
+    class MiniCim(numberOfMotors: Int): Motor() {
         override val freeSpeed = rpmToRad(6200.0)
         override val freeCurrent = 1.5
         override val stallTorque = 1.4 * numberOfMotors
